+++ config/global.ini.php
@@ -1012,7 +1012,7 @@ Plugins[] = Intl
 Plugins[] = Marketplace
 Plugins[] = ProfessionalServices
 Plugins[] = UserId
-Plugins[] = CustomPiwikJs
+Plugins[] = CustomTrackerJs
 Plugins[] = Tour
 
 [PluginsInstalled]